update header section

+++ src/app/seller-home/seller-home.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { product } from '../data-type';
 import { ProductService } from '../services/product.service';
 import { faTrash, faEdit } from '@fortawesome/free-solid-svg-icons';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-seller-home',
   templateUrl: './seller-home.component.html',
@@ -30,18 +31,59 @@ pageSize = 5; // Number of items per page
 
   }
 
-  deleteProduct(id: number) {
-    this.product.deleteProduct(id).subscribe((result) => {
-      if (result) {
-        this.productMessage = 'Product is deleted';
-
-        this.list();
-      }
-    });
-    setTimeout(() => {
-      this.productMessage = undefined;
-    }, 3000);
+      show() {
+ 
   }
+
+
+ deleteProduct(id: number) {   
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: "btn btn-success",
+      cancelButton: "btn btn-danger"
+    },
+    buttonsStyling: false
+  });
+
+  swalWithBootstrapButtons.fire({
+    title: "Are you sure?",
+    text: "You won't be able to revert this!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonText: "Yes, delete it!",
+    cancelButtonText: "No, cancel!",
+    reverseButtons: true
+  }).then((result) => {
+    if (result.isConfirmed) {
+      this.product.deleteProduct(id).subscribe((result) => {
+        if (result) {
+          this.productMessage = 'Product is deleted';
+          swalWithBootstrapButtons.fire(
+            "Deleted!",
+            "Your product has been deleted.",
+            "success"
+          );
+          this.list(); // Refresh the product list
+        }
+      });
+    } else if (result.dismiss === Swal.DismissReason.cancel) {
+      swalWithBootstrapButtons.fire(
+        "Cancelled",
+        "Your product is safe :)",
+        "error"
+      );
+    }
+  });
+
+  setTimeout(() => {
+    this.productMessage = undefined;
+  }, 3000);
+}
+
+
+
+
+
 
   list() {
     this.product.productList().subscribe((result) => {
@@ -71,4 +113,8 @@ pageSize = 5; // Number of items per page
     const endIndex = startIndex + this.pageSize;
     this.paginatedData = this.productList.slice(startIndex, endIndex);
   }
+}
+
+function show() {
+  throw new Error('Function not implemented.');
 }
